requirements.txt
autopep8==1.5.4
chardet==3.0.4
click==7.1.2
fasttext==0.9.2
joblib==0.17.0
langdetect==1.0.8
langid==1.1.6
nltk==3.5
numpy==1.19.2
pybind11==2.5.0
pycld3==0.20
pycodestyle==2.6.0
regex==2020.10.11
scipy==1.5.2
six==1.15.0
textblob==0.15.3
toml==0.10.1
tqdm==4.50.2
